refactor: extract MDX rendering to client component with CSP nonce support
